refactor: split up data collection by OS
