+++ Cargo.toml
@@ -42,6 +42,7 @@ chrono = "0.4.19"
 crossterm = "0.18.2"
 ctrlc = { version = "3.1.9", features = ["termination"] }
 clap = "2.33"
+cfg-if = "1.0"
 dirs-next = "2.0.0"
 futures = "0.3.14"
 futures-timer = "3.0.2"
